fix tests_ecg()
@johndoenut it seems like the ecg signal simulated in travis gives non-trivial different ecg signal than the one simulated in our machines (even with the random_state). It could possibly because of differences in hardware? I'm not too sure. Nevertheless, in order for the `test_ecg_rate()` to pass in travis, I previously had to increase the atol of `np.allclose(rate.mean(), 81)` to 2.  Hence, I'm reverting your changes. But if you have any better suggestion, do let me know!

For the assertion of `rpeaks_corrected["ECG_R_Peaks"].sum(dtype=np.int64)` in `test_ecg_fixpeaks()`, the sum generated in travis seems to be vastly different with the number specified. Hence, I'm trying to simulate the ecg signal by using a different method (simple) and let's see how it fares in the test.