Optimize _ecg_findpeaks_MWA
This optimization speeds up the moving average computation by an order of
magnitude. This is the most time-consuming part of pantompkins peak
detection, which after this optimization is about 5x faster on my computer.

Note the TODO about the _ecg_findpeaks_MWA output. I intentionally kept
this as a pure optimization with no impact on the output of the function,
but I do wonder whether the earlier implementation was actually incorrect,
or if the output is intentionally shifted by one.