Vectorized peak detection in _ecg_findpeaks_peakdetect
It's far faster to use scipy.signal.find_peaks than to incrementally
build the peaks array by stepping through each sample in Python code.
And using enumerate() frees us from having to separately keep track of
the index of the current peak.

Note the TODO I left about the plateau_size=(1, 1) setting I had to
include to prevent detection of flat peaks. It's not obvious why such
peaks should not be detected, but for this commit I wanted to maintain
the exact same behavior of the original py-ecg-detectors code.