Drop the indexes array in _ecg_findpeaks_peakdetect
Only the last two elements of that array were being accessed. One of those
was already available as the index variable and the other one was easy to
keep track of as a new last_index variable.

Also add a similar last_peak variable to reduce the number of signal_peaks
array lookups, and remove the now unnecessary signal_peaks sentinel value.