Fix rare peak-inside-padding issues in _ecg_findpeaks_kalidas
I occasionally see failures in `kalidas2017` peak detection caused by the algorithm returning peak indices that exceed the length of the input signal.

The cause of this problem is the padding needed by the SWT computation, and the simple fix is to drop the padding after the SWT computation is done.

I can't share the data that triggers this problem, and I couldn't think of an easy way to generate an alternative test case. Thus unfortunately I don't have a test case to go with this change. At least the change itself is very simple, and I've verified it locally with the data I have.