Update eda_plot.py
Fix a bug where you had an exception if there was not the same number of peaks and onsets.
This can happen on signal that has been cut (ex: depending on difference conditions).
The fix just ignores the first peak (only for display) when there is no onset for the first peak.